doc_analyzer_mcp_server/requirements.txt
mcp[cli]
pandas
python-docx
